backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0

# Database drivers
psycopg2-binary>=2.9.0  # PostgreSQL
mysqlclient>=2.2.0  # MySQL (alternative: PyMySQL>=1.1.0)